#43  Move helpers outside CqlParser
